requirements.txt
flask
google-generativeai
reportlab
python-dotenv
pymongo
gunicorn